identify issues and update doc generation

+++ src/battery.zig
@@ -1,5 +1,7 @@
+//! This file was derived from the 'battery' symbol group 
+// This file was programmatically generated. These symbols may resolve to the incorrect location if firmware changes.
 const types = @import("./types.zig");
-pub extern const vexBatteryVoltageGet: *const fn() callconv(.c) i32; // 0xa00 
-pub extern const vexBatteryCurrentGet: *const fn() callconv(.c) i32; // 0xa04 
-pub extern const vexBatteryTemperatureGet: *const fn() callconv(.c) i32; // 0xa08 
-pub extern const vexBatteryCapacityGet: *const fn() callconv(.c) i32; // 0xa0c 
+pub extern const vexBatteryVoltageGet: *const fn() callconv(.c) i32;
+pub extern const vexBatteryCurrentGet: *const fn() callconv(.c) i32;
+pub extern const vexBatteryTemperatureGet: *const fn() callconv(.c) f64;
+pub extern const vexBatteryCapacityGet: *const fn() callconv(.c) f64;
